Replacing mmController, introspectionSystem, storageSystem, netSystem and tcpSystem with VeinManager.
VeinManager contains all basic Vein functions. Entitie information are
provided by the _VEIN entity. _VEIN is a part of the former _System
entity. This commit removes the _System entity. It will be reintroduced in a
future commit.

The VeinManager subsystem has to be the first appended Subsystem. The
order of any other subsystem is not important .

Signed-off-by: bhamacher <b.hamacher@zera.de>